requirements.txt
pyrogram==1.4.12
py-tgcalls==0.8.6 
TgCrypto
python-dotenv
youtube_dl
youtube_search_python 
requests
aiohttp[speedups]
aiofiles
asyncio==3.4.3
youtube_search
python-arq
search_engine_parser
wget 
ffmpeg-python
Pillow==9.1.0
ujson
pymongo[srv]
psutil
motor
speedtest-cli
lyricsgenius
yt_dlp
yt-dlp
